If the pivotal story ID is already in the git commit message, don't add it again.

+++ lib/pivotal-integration/command/prepare-commit-msg.sh
@@ -16,11 +16,17 @@
 
 CURRENT_BRANCH=$(git branch | grep "*" | sed "s/* //")
 STORY_ID=$(git config branch.$CURRENT_BRANCH.pivotal-story-id)
+STORY_TAG="[#$STORY_ID]"
 
 if [[ $2 != "commit" && -n $STORY_ID ]]; then
 	ORIG_MSG_FILE="$1"
 	TEMP=$(mktemp /tmp/git-XXXXX)
 
-  (cat "$1" ; printf "\n\n[#$STORY_ID]") > "$TEMP"
+  # don't duplicate the tag if it's already there
+  if grep -F "$STORY_TAG" "$ORIG_MSG_FILE"; then
+    exit 0
+  fi
+
+	(printf "\n\n$STORY_TAG" ; cat "$1") > "$TEMP"
 	cat "$TEMP" > "$ORIG_MSG_FILE"
 fi
